WIP. 3-emd not in working status
