+++ .gitignore
@@ -102,3 +102,11 @@ Workflow/output/2-umap/4%-1.txt__final_umap_test.txt
 Workflow/output/2-umap/4%-2.txt__final_umap_test.txt
 Workflow/output/1-data_preprocess/Testing/figure-1_full-gate_1st-cytobank-export_no-arcsinh_TRUNCATED.txt
 Workflow/output/2-umap/merged_final_umap_test.txt
+Workflow/4-emd/3-emd_no-norm_arcsinh_wo-plots.py
+Workflow/4-emd/emd_no-norm_arcsinh_wo-plots.py
+Workflow/4-emd/Untitled.ipynb
+Workflow/4-emd/.ipynb_checkpoints/Untitled-checkpoint.ipynb
+Workflow/input/3-emd/3%-1.txt
+Workflow/input/3-emd/3%-2.txt
+Workflow/input/3-emd/4%-1.txt
+Workflow/input/3-emd/4%-2.txt
