+++ Workflow/aux_functions.py
@@ -5,9 +5,6 @@ import umap
 import sys
 import os
 
-import warnings
-warnings.filterwarnings('ignore')
-
 #Function to concatenate all files
 def concatenate_fcs(folder_name):
     input_files = [f for f in os.listdir(f"./{folder_name}") if f.endswith(".txt")]
@@ -22,6 +19,16 @@ def concatenate_fcs(folder_name):
         df["Cell_Index"] = df["Cell_Index"].apply(lambda x: str(fcounter)+"-"+str(x)) #File+ID
         no_arc = no_arc.append(df, ignore_index=True)
     return no_arc, input_files
+
+#Arcsinh transform the data
+def arcsinh_transf(cofactor, no_arc):
+    #Select only the columns containing the markers (as they start with a number for the isotope)
+    cols = [x for x in no_arc.columns if x[0].isdigit()]
+    #Apply the arcsinh only to those columns (don't want to change time or any other)
+    arc = no_arc.apply(lambda x: np.arcsinh(x/cofactor) if x.name in cols else x)
+    # put back the 'file_origin' column to the arcsinh-transformed data
+    arc["file_origin"] = no_arc["file_origin"]
+    return arc, cols
 
 # Random downsampling of a dataframe to n rows
 def downsample_df(df, n):
